refactor: transfer match maker logic to repo
