+++ lib/app/view/app.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:game_client/game_client.dart';
 import 'package:go_router/go_router.dart';
+import 'package:match_maker_repository/match_maker_repository.dart';
 import 'package:provider/provider.dart';
 import 'package:top_dash/app_lifecycle/app_lifecycle.dart';
 import 'package:top_dash/audio/audio_controller.dart';
 import 'package:top_dash/l10n/l10n.dart';
-import 'package:top_dash/match_making/match_maker.dart';
 import 'package:top_dash/router/router.dart';
 import 'package:top_dash/settings/persistence/persistence.dart';
 import 'package:top_dash/settings/settings.dart';
@@ -32,7 +32,7 @@ class App extends StatefulWidget {
   const App({
     required this.settingsPersistence,
     required this.gameClient,
-    required this.matchMaker,
+    required this.matchMakerRepository,
     this.router,
     super.key,
   });
@@ -41,7 +41,7 @@ class App extends StatefulWidget {
 
   final GameClient gameClient;
 
-  final MatchMaker matchMaker;
+  final MatchMakerRepository matchMakerRepository;
 
   final GoRouter? router;
 
@@ -58,7 +58,7 @@ class _AppState extends State<App> {
       child: MultiProvider(
         providers: [
           Provider.value(value: widget.gameClient),
-          Provider.value(value: widget.matchMaker),
+          Provider.value(value: widget.matchMakerRepository),
           Provider<SettingsController>(
             lazy: false,
             create: (context) => SettingsController(
