+++ lib/match_making/match_making.dart
@@ -1,3 +1,2 @@
 export 'bloc/match_making_bloc.dart';
-export 'match_maker.dart';
 export 'views/views.dart';
