+++ lib/match_making/views/match_making_page.dart
@@ -2,6 +2,7 @@ import 'package:flutter/widgets.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:game_client/game_client.dart';
 import 'package:go_router/go_router.dart';
+import 'package:match_maker_repository/match_maker_repository.dart';
 import 'package:top_dash/match_making/match_making.dart';
 
 class MatchMakingPage extends StatelessWidget {
@@ -23,10 +24,10 @@ class MatchMakingPage extends StatelessWidget {
   Widget build(BuildContext context) {
     return BlocProvider(
       create: (context) {
-        final matchMaker = context.read<MatchMaker>();
+        final matchMakerRepository = context.read<MatchMakerRepository>();
         final gameClient = context.read<GameClient>();
         return MatchMakingBloc(
-          matchMaker: matchMaker,
+          matchMakerRepository: matchMakerRepository,
           gameClient: gameClient,
           cardIds: playerCardIds,
         )..add(
