+++ packages/match_maker_repository/lib/match_maker_repository.dart
@@ -2,3 +2,4 @@
 library match_maker_repository;
 
 export 'src/match_maker_repository.dart';
+export 'src/models/models.dart';
