+++ packages/match_maker_repository/lib/src/models/models.dart
@@ -1,0 +1,1 @@
+export 'match.dart';
