+++ pubspec.yaml
@@ -28,6 +28,8 @@ dependencies:
   go_router: ^6.2.0
   intl: ^0.17.0
   logging: ^1.1.1
+  match_maker_repository:
+    path: packages/match_maker_repository
   provider: ^6.0.5
   shared_preferences: ^2.0.17
