+++ test/match_making/bloc/match_making_state_test.dart
@@ -2,6 +2,7 @@
 
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:match_maker_repository/match_maker_repository.dart';
 import 'package:top_dash/match_making/match_making.dart';
 
 void main() {
